Moved some results classes around
