+++ Plan/common/src/main/java/com/djrapitops/plan/extension/implementation/results/ExtensionInformation.java
@@ -14,7 +14,7 @@
  *  You should have received a copy of the GNU Lesser General Public License
  *  along with Plan. If not, see <https://www.gnu.org/licenses/>.
  */
-package com.djrapitops.plan.extension.implementation.results.player;
+package com.djrapitops.plan.extension.implementation.results;
 
 import com.djrapitops.plan.extension.icon.Icon;
