+++ Plan/common/src/test/java/com/djrapitops/plan/db/CommonDBTest.java
@@ -47,10 +47,10 @@ import com.djrapitops.plan.db.patches.Patch;
 import com.djrapitops.plan.extension.DataExtension;
 import com.djrapitops.plan.extension.ExtensionServiceImplementation;
 import com.djrapitops.plan.extension.annotation.*;
-import com.djrapitops.plan.extension.implementation.results.player.ExtensionBooleanData;
+import com.djrapitops.plan.extension.implementation.results.ExtensionBooleanData;
+import com.djrapitops.plan.extension.implementation.results.ExtensionStringData;
+import com.djrapitops.plan.extension.implementation.results.ExtensionTabData;
 import com.djrapitops.plan.extension.implementation.results.player.ExtensionPlayerData;
-import com.djrapitops.plan.extension.implementation.results.player.ExtensionStringData;
-import com.djrapitops.plan.extension.implementation.results.player.ExtensionTabData;
 import com.djrapitops.plan.extension.implementation.storage.queries.ExtensionPlayerDataQuery;
 import com.djrapitops.plan.extension.implementation.storage.transactions.results.RemoveUnsatisfiedConditionalResultsTransaction;
 import com.djrapitops.plan.system.PlanSystem;
